requirements.txt
altgraph==0.17.4
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
flatbuffers==24.3.25
humanfriendly==10.0
idna==3.6
imageio==2.34.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy_loader==0.3
llvmlite==0.42.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
onnxruntime==1.17.1
opencv-python-headless==4.9.0.80
packaging==24.0
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==5.26.0
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyMatting==1.1.12
pyreadline3==3.4.1
pywin32-ctypes==0.2.2
referencing==0.34.0
rembg==2.0.56
requests==2.31.0
rpds-py==0.18.0
scikit-image==0.22.0
scipy==1.12.0
setuptools==69.2.0
sympy==1.12
tifffile==2024.2.12
tqdm==4.66.2
urllib3==2.2.1
